+++ creme/metrics/__init__.py
@@ -17,7 +17,6 @@ from .mae import MAE
 from .mcc import MCC
 from .mse import MSE
 from .multioutput import RegressionMultiOutput
-from .per_class import PerClass
 from .precision import MacroPrecision
 from .precision import MicroPrecision
 from .precision import Precision
@@ -53,7 +52,6 @@ __all__ = [
     'MicroRecall',
     'MSE',
     'MultiFBeta',
-    'PerClass',
     'Precision',
     'Recall',
     'RegressionMultiOutput',